refactored code to make it more readable.
